third_party/WebKit/LayoutTests/fast/workers/storage/change-version-sync-expected.txt

CONSOLE WARNING: 'openDatabaseSync' is deprecated. Please switch to Indexed Database API.
CONSOLE WARNING: 'openDatabaseSync' is deprecated. Please switch to Indexed Database API.
Test that version updates in sync databases work correctly.
PASS: db.version is 3 as expected.
